Remove ToolShortcutLinksView Set shortcut items on app enter background
